move functions out of window
