+++ kwm/helpers.h
@@ -90,4 +90,36 @@ ConvertHexRGBAToColor(unsigned int Color)
     return Result;
 }
 
+inline CGPoint
+GetCursorPos()
+{
+    CGEventRef Event = CGEventCreate(NULL);
+    CGPoint Cursor = CGEventGetLocation(Event);
+    CFRelease(Event);
+
+    return Cursor;
+}
+
+inline std::string
+GetUTF8String(CFStringRef Temp)
+{
+    std::string Result;
+
+    if(!CFStringGetCStringPtr(Temp, kCFStringEncodingUTF8))
+    {
+        CFIndex Length = CFStringGetLength(Temp);
+        CFIndex Bytes = 4 * Length + 1;
+        char *TempUTF8StringPtr = (char*) malloc(Bytes);
+
+        CFStringGetCString(Temp, TempUTF8StringPtr, Bytes, kCFStringEncodingUTF8);
+        if(TempUTF8StringPtr)
+        {
+            Result = TempUTF8StringPtr;
+            free(TempUTF8StringPtr);
+        }
+    }
+
+    return Result;
+}
+
 #endif
